+++ lib/src/soloud.dart
@@ -5,6 +5,7 @@ import 'dart:ffi' as ffi;
 import 'dart:isolate';
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_soloud/src/audio_isolate.dart';
 import 'package:flutter_soloud/src/audio_source.dart';
@@ -557,6 +558,46 @@ interface class SoLoud {
 
     completer.complete(true);
     return completer.future;
+  }
+
+  /// Stops the engine and disposes of all resources, including sounds
+  /// and the audio isolate in an synchronous way.
+  /// 
+  /// This method is meant to be called when exiting the app. For example
+  /// within the `dispose()` of the uppermost widget in the tree
+  /// or inside [AppLifecycleListener.onExitRequested].
+  /// 
+  /// During the normal app life cycle and if you want to shutdown the player,
+  /// please use [shutdown] which safer and it is meant to throw errors.
+  void deinit() {
+    _log.finest('deinit() called');
+    /// check if we are in the middle of an initialization.
+    if (_initializeCompleter != null) {
+      _log.warning('deinit() called while already initializing.');
+      _initializeCompleter = null;
+    }
+
+    /// if not already initialized, just return.
+    if (!_isInitialized) {
+      // The engine isn't initialized.
+      _log.warning('deinit() called when the engine is not initialized');
+      return;
+    }
+
+    /// reset broadcast and kill isolate
+    _isolateToMainStream?.close();
+    _isolateToMainStream = null;
+    _mainToIsolateStream = null;
+    _returnedEvent?.close();
+    _returnedEvent = null;
+    _isolate?.kill();
+    _isolate = null;
+    _isEngineInitialized = false;
+    _isInitialized = false;
+
+    SoLoudController().soLoudFFI.disposeAllSound();
+    SoLoudController().soLoudFFI.deinit();
+    _activeSounds.clear();
   }
 
   /// return true if the audio isolate is running
@@ -1361,6 +1402,83 @@ interface class SoLoud {
       throw const SoLoudNotInitializedException();
     }
     return SoLoudController().soLoudFFI.getIsValidVoiceHandle(handle);
+  }
+
+  /// Returns the number of concurrent sounds that are playing at the moment.
+  int getActiveVoiceCount() {
+    if (!isInitialized) {
+      throw const SoLoudNotInitializedException();
+    }
+    return SoLoudController().soLoudFFI.getActiveVoiceCount();
+  }
+
+  /// Returns the number of concurrent sounds that are playing a
+  /// specific audio source.
+  int getCountAudioSource(int soundHash) {
+    if (!isInitialized) {
+      throw const SoLoudNotInitializedException();
+    }
+    return SoLoudController().soLoudFFI.getCountAudioSource(soundHash);
+  }
+
+  /// Returns the number of voices the application has told SoLoud to play.
+  int getVoiceCount() {
+    if (!isInitialized) {
+      throw const SoLoudNotInitializedException();
+    }
+    return SoLoudController().soLoudFFI.getVoiceCount();
+  }
+
+  // /// Get a sound's protection state.
+  // bool getProtectVoice(int handle) {
+  //   if (!isInitialized) {
+  //     throw const SoLoudNotInitializedException();
+  //   }
+  //   return SoLoudController().soLoudFFI.getProtectVoice(handle);
+  // }
+
+  // /// Set a sound's protection state.
+  // ///
+  // /// Normally, if you try to play more sounds than there are voices,
+  // /// SoLoud will kill off the oldest playing sound to make room.
+  // /// This will most likely be your background music. This can be worked
+  // /// around by protecting the sound.
+  // /// If all voices are protected, the result will be undefined.
+  // ///
+  // /// [handle]  handle to check.
+  // /// [protect] whether to protect or not.
+  // void setProtectVoice(int handle, bool protect) {
+  //   if (!isInitialized) {
+  //     throw const SoLoudNotInitializedException();
+  //   }
+  //   SoLoudController().soLoudFFI.setProtectVoice(handle, protect);
+  // }
+
+  /// Get the current maximum active voice count.
+  int getMaxActiveVoiceCount() {
+    if (!isInitialized) {
+      throw const SoLoudNotInitializedException();
+    }
+    return SoLoudController().soLoudFFI.getMaxActiveVoiceCount();
+  }
+
+  /// Set the current maximum active voice count.
+  /// If voice count is higher than the maximum active voice count,
+  /// SoLoud will pick the ones with the highest volume to actually play.
+  /// [maxVoiceCount] the max concurrent sounds that can be played.
+  ///
+  /// NOTE: The number of concurrent voices is limited, as having unlimited
+  /// voices would cause performance issues, as well as lead to unnecessary
+  /// clipping. The default number of concurrent voices is 16, but this can be
+  /// adjusted at runtime. The hard maximum number is 4095, but if more are
+  /// required, SoLoud can be modified to support more. But seriously, if you
+  /// need more than 4095 sounds at once, you're probably going to make
+  /// some serious changes in any case.
+  void setMaxActiveVoiceCount(int maxVoiceCount) {
+    if (!isInitialized) {
+      throw const SoLoudNotInitializedException();
+    }
+    SoLoudController().soLoudFFI.setMaxActiveVoiceCount(maxVoiceCount);
   }
 
   /// Return a floats matrix of 256x512.